Bidirectional jet added
The bidirectional jet involves a spherical region in the middle of the grid at (0,0,0) with a radius OSPH (new input parameter) and a mirrored (actually, a rotationally symmetric) nozzle region within. The outflow radius ORAD, retains it's original meaning - it is the radius at the exit point of the flow through the nozzle.

Tilt and precession were tested with the bidirectional jet.

Bug fixes for parallel jet.