Bugfixes, espcially regarding angular velocities.
In many parts of the code, e.g. clouds.c where the Keplerian velocit is set, the poloidal velocity was given as a linear velocity, rather than an angular velocity.
In particular, all the macros in macros_usr were changed to to be the ones that divide or multiply correctly by a radius. I hope this does not make the code crash in cases where conversions to or from velocities in spherical were made. Most (all?) vector conversions where in terms of radial velocities, so nothing might be affected, but this certainly needs more testing.