Equilateral 2D spherical polars
New overwritten file runtime_setup, in which the hack appears.
In spherical >1D,  if patch_npoints of a segment is negative, then the number indicates a reference patch of the IDIR grid, which should be used to create (roughly) equilateral cells by calculating the appropriate npoints for the JDIR direction of the current patch. (Note: "segment" = "patch").